Playlist view is now 3rd-party agnostic, initializing player based on chosen music service

+++ src/components/_playlist/playlist.controller.js
@@ -1,15 +1,25 @@
+import _ from 'lodash';
+
 export default class PlaylistController {
 	/* @ngInject */
-	constructor(breadcrumbService, playerWidgetService) {
-		breadcrumbService.buildPlaylistBreadcrumb(this.servicePlaylist);
+	constructor(breadcrumbService, playerWidgetService, serviceDropdown) {
+		breadcrumbService.buildPlaylistBreadcrumb(this.localPlaylist);
 		this.playerWidgetService = playerWidgetService;
+		this.serviceDropdown = serviceDropdown;
 	}
 	showPlayer() {
+		let service = this.serviceDropdown.service.name;
+		let servicePlaylist = _.find(this.localPlaylist.external_playlists, {
+			source: service
+		});
+		if (!servicePlaylist) {
+			//@TODO: handle not found 3rd party playlist here
+			return;
+		}
 		this.playerWidgetService.destroy();
 		this.playerWidgetService.player = {
-			service: 'deezer',
-			tracks: this.servicePlaylist.tracks.data,
-			servicePlaylistId: this.servicePlaylist.id,
+			service,
+			servicePlaylistId: servicePlaylist.service_playlist_id,
 			localPlaylistId: this.localPlaylist.id,
 			show: true
 		};
